Split ROI sampling into two parts: Generating ROIs and sampling ROIs
ROIs are generated from a mask such that the center of the ROI is inside
the mask and the entire ROI is contained in the image. This process generates
a csv file of ROIs defined by their start index and size.

ROIs are sampled from an image by reading each ROI from the csv file and then
sampling that region of the image.